Revert "Document and condense AtTabInfo"
This reverts commit c747d9ddd7052ffc712fea52b2e7430ff581b835.
