+++ AtEventDisplay/AtEventTab/AtTabInfoFairRoot.cxx
@@ -1,0 +1,1 @@
+#include "AtTabInfoFairRoot.h"
