+++ AtEventDisplay/AtEventTab/AtTabMain.cxx
@@ -7,6 +7,8 @@
 #include "AtPad.h" // for AtPad
 #include "AtRawEvent.h"
 #include "AtTabInfo.h"
+#include "AtTabInfoBase.h"
+#include "AtTabInfoFairRoot.h"
 
 #include <FairLogger.h> // for LOG
