+++ AtEventDisplay/CMakeLists.txt
@@ -17,6 +17,8 @@ AtEventTab/AtTabMacro.cxx
 AtEventTab/AtTabBase.cxx
 
 AtEventTab/AtTabInfo.cxx
+AtEventTab/AtTabInfoFairRoot.cxx
+AtEventTab/AtTabInfoBase.cxx
 
 AtEventManager.cxx
 AtEventManagerNew.cxx
